requirements.txt
dj-database-url==0.3.0
Django==1.9.6
django-debug-toolbar==1.3.2
django-postgrespool==0.3.0
django-zurb-foundation==5.5.0
django-webpack-loader==0.2.4
gunicorn==19.3.0
psycopg2==2.6
SQLAlchemy==1.0.4
sqlparse==0.1.16
wheel==0.24.0
whitenoise==1.0.6
django-extra-views==0.7.1
jsonfield==1.0.3
djangorestframework==3.3.3
